pyproject.toml
[tool.poetry]
name = "simple-tg-bot-exchange-rates"
version = "0.1.0"
description = ""
authors = ["rizzzon <rizzzon116@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "13.13"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
